+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -5,8 +5,8 @@ import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.entity.ReservationThemeEntity;
 import roomescape.theme.repository.ReservationThemeRepository;
 import roomescape.time.repository.ReservationTimeRepository;
-import roomescape.reservation.dto.ReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.dto.request.ReservationRequest;
+import roomescape.reservation.service.dto.response.ReservationResponse;
 import roomescape.reservation.entity.ReservationEntity;
 import roomescape.time.entity.ReservationTimeEntity;
 import roomescape.exception.BadRequestException;
