+++ src/main/java/roomescape/theme/service/dto/request/ReservationThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.service.dto.request;
 
 import roomescape.theme.entity.ReservationThemeEntity;
