+++ src/main/java/roomescape/theme/service/dto/response/ReservationThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.service.dto.response;
 
 import roomescape.theme.entity.ReservationThemeEntity;
