+++ src/main/java/roomescape/time/controller/user/UserReservationTimeController.java
@@ -3,7 +3,7 @@ package roomescape.time.controller.user;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.time.dto.AvailableReservationTimeResponse;
+import roomescape.time.service.dto.response.AvailableReservationTimeResponse;
 import roomescape.time.service.ReservationTimeService;
 
 import java.time.LocalDate;
