--- src/main/java/roomescape/time/dto/AvailableReservationTimeResponse.java
@@ -1,8 +1,0 @@
-package roomescape.time.dto;
-
-public record AvailableReservationTimeResponse(
-        Long id,
-        String startAt,
-        boolean alreadyBooked
-) {
-}
