+++ src/main/java/roomescape/time/service/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.time.dto;
+package roomescape.time.service.dto.response;
 
 import roomescape.time.entity.ReservationTimeEntity;
