+++ src/test/java/roomescape/MissionStepTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.reservation.controller.ReservationController;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.dto.response.ReservationResponse;
 
 import java.lang.reflect.Field;
 import java.sql.Connection;
